2019.4.6 添加功能点，角色的搜索功能 在查询展示和分配功能点，角色时 均可以搜索和分页

+++ permission-server/src/main/java/com/clsaa/dop/server/permission/service/PermissionService.java
@@ -46,6 +46,14 @@ public class PermissionService {
     //关联关系service
     private UserRoleMappingService userRoleMappingService;
 
+    @Autowired
+    //用户数据
+    private UserDataService userDataService;
+
+    @Autowired
+    //用户数据规则
+    private UserRuleService userRuleService;
+
 /* *
  *
  *  * @param id 功能点ID
@@ -67,21 +75,28 @@ public class PermissionService {
     public void createPermission(Long parentId,String name,Integer isPrivate,String description,
                                        Long cuser,Long muser)
     {
+        if(checkUserPermission("创建功能点",cuser))
+        {
+            Permission existPermission=this.permissionRepository.findByName(name);
+            BizAssert.allowed(existPermission==null, BizCodes.REPETITIVE_PERMISSION_NAME);
+            Permission permission= Permission.builder()
+                    .parentId(parentId)
+                    .name(name)
+                    .isPrivate(isPrivate)
+                    .description(description)
+                    .cuser(cuser)
+                    .muser(muser)
+                    .ctime(LocalDateTime.now())
+                    .mtime(LocalDateTime.now())
+                    .deleted(false)
+                    .build();
+            permissionRepository.saveAndFlush(permission);
+            userDataService.addData(
+                    userRuleService.findUniqueRule("in","permissionId",
+                            roleService.findByName("权限管理员").getId()).getId(),
+                    cuser,permission.getId(),cuser,muser);
+        }
 
-        Permission existPermission=this.permissionRepository.findByName(name);
-        BizAssert.allowed(existPermission==null, BizCodes.REPETITIVE_PERMISSION_NAME);
-        Permission permission= Permission.builder()
-                .parentId(parentId)
-                .name(name)
-                .isPrivate(isPrivate)
-                .description(description)
-                .cuser(cuser)
-                .muser(muser)
-                .ctime(LocalDateTime.now())
-                .mtime(LocalDateTime.now())
-                .deleted(false)
-                .build();
-         permissionRepository.saveAndFlush(permission);
 
     }
 
@@ -96,25 +111,46 @@ public class PermissionService {
         return null;
     }
     //分页查询所有功能点
-    public Pagination<PermissionV1> getPermissionV1Pagination(Integer pageNo, Integer pageSize)
+    public Pagination<PermissionV1> getPermissionV1Pagination(Integer pageNo, Integer pageSize,Long userId,String key)
     {
         Sort sort = new Sort(Sort.Direction.DESC, "mtime");
-        int count = (int) this.permissionRepository.count();
+        int count = 0;
 
         Pagination<PermissionV1> pagination = new Pagination<>();
         pagination.setPageNo(pageNo);
         pagination.setPageSize(pageSize);
-        pagination.setTotalCount(count);
 
+        List<Permission> permissionList=new ArrayList<>();
+        //未填写搜索关键字，则查询全部
+        if(key.equals(""))
+        {
+            permissionList = this.permissionRepository.findAll(sort);
+        }
+        //填写了搜索关键字，带条件查询
+        else {
+            permissionList = this.permissionRepository.findByNameLike("%"+key+"%");
+        }
+
+        List<Long> idList=userDataService.findAllIds("查询功能点",userId,"permissionId");
+
+        List<Permission> permissionList1=new ArrayList<>();
+        for(Permission permission :permissionList)
+        {
+            for(Long id :idList)
+            {
+                if(permission.getId()==id)
+                {permissionList1.add(permission);count++;}
+            }
+        }
+        pagination.setTotalCount(count);
         if (count == 0) {
             pagination.setPageList(Collections.emptyList());
             return pagination;
         }
 
-        Pageable pageRequest = PageRequest.of(pagination.getPageNo() - 1, pagination.getPageSize(), sort);
-        List<Permission> permissionList = this.permissionRepository.findAll(pageRequest).getContent();
-        pagination.setPageList(permissionList.stream().map(p -> BeanUtils.convertType(p, PermissionV1.class)).collect(Collectors.toList()));
+        permissionList1=permissionList1.subList((pageNo-1)*pageSize, (pageNo*pageSize<count)? pageNo*pageSize:count);
 
+        pagination.setPageList(permissionList1.stream().map(p -> BeanUtils.convertType(p, PermissionV1.class)).collect(Collectors.toList()));
         return pagination;
     }
 
@@ -126,11 +162,16 @@ public class PermissionService {
 
     //根据ID删除功能点,并删除关联关系
     @Transactional
-    public void deleteById(Long id)
+    public void deleteById(Long id,Long userId)
     {
-        rolePermissionMappingService.deleteByPermissionId(id);
-        permissionRepository.deleteById(id);
-
+        if(checkUserPermission("删除功能点",userId))
+        {
+            if(userDataService.check("删除功能点",userId,"permissionId",id))
+            {
+                rolePermissionMappingService.deleteByPermissionId(id);
+                permissionRepository.deleteById(id);
+            }
+        }
     }
 
     //创建或编辑角色时，需要勾选该角色对应的功能点，所以要返回全部功能点
